页面开发

+++ admin/ReviewService.php
@@ -8,9 +8,9 @@ class ReviewService
     {
         $sql = "select a.id,a.pageid,a.title,a.viewnum,a.cover,a.imgnum,a.disnum,a.praisenum,a.showindex,a.status,a.cuserid,a.cusername,
                 a.cdate,a.udate
-                from page a  where 1=1 " ;
-        if($status!=-1){
-            $sql .= " and status=" .$status;
+                from page a  where 1=1 ";
+        if ($status != -1) {
+            $sql .= " and status=" . $status;
         }
         if (!empty($keyword)) {
             $sql .= " and a.title like '%" . $keyword . "%'";
@@ -22,10 +22,10 @@ class ReviewService
         if (empty($res)) return $pages;
         foreach ($res as $data) {
             $pageid = $data['pageid'];
-            $sqlImg="select id imgid,img from page_img where pageid=".$pageid;
+            $sqlImg = "select id imgid,img from page_img where pageid=" . $pageid;
             $imgs = $mysql->executeReturnObj($sqlImg);
             $data['imgs'] = $imgs;
-            $pages[]=$data;
+            $pages[] = $data;
         }
         $mysql->closeCon();
         return $pages;
@@ -34,8 +34,8 @@ class ReviewService
     public static function getTotal($keyword, $status)
     {
         $where = " 1=1 ";
-        if($status!=-1){
-            $where .= " and status=" .$status;
+        if ($status != -1) {
+            $where .= " and status=" . $status;
         }
         if (!empty($keyword)) {
             $where .= " and title like '%" . $keyword . "%'";
@@ -46,7 +46,7 @@ class ReviewService
         return $total;
     }
 
-    public static function delImg($imgid)
+    public static function delImg($imgid, $deltype = 1)
     {
         $mysql = new MySQL();
         $sqlImg = "select * from page_img where id=" . $imgid;
@@ -68,10 +68,13 @@ class ReviewService
         $mysql->execute($sqlDel);
         //删除文件
         $fullPath = MIMGPATH . $img;
-
-        if(file_exists($fullPath)){
+        if (file_exists($fullPath) && $deltype == 1) {
             unlink($fullPath);
         }
+        //更新图片数
+        $upNumSql = "update page set imgnum=imgnum-1 where pageid=" . $pageid;
+        $mysql->execute($upNumSql);
+
         if ($cover == $img) {
             $sqlImg2 = "select * from page_img where pageid=" . $pageid;
             $imgNw = $mysql->executeReturnFirstObj($sqlImg2);
@@ -104,18 +107,18 @@ class ReviewService
         return true;
     }
 
-    public static function delPage($pageid,$deltype=1)
+    public static function delPage($pageid, $deltype = 1)
     {
         $mysql = new MySQL();
         //删除图片和图片数据
         $sqlImg = "select * from page_img where pageid=" . $pageid;
         $imgs = $mysql->executeReturnObj($sqlImg);
-        if (!empty($imgs) && 1==$deltype) {
+        if (!empty($imgs) && 1 == $deltype) {
             foreach ($imgs as $img) {
                 $imgName = $img['img'];
                 //删除文件
                 $fullPath = MIMGPATH . $imgName;
-                if(file_exists($fullPath)){
+                if (file_exists($fullPath)) {
                     unlink($fullPath);
                 }
             }
@@ -139,8 +142,8 @@ class ReviewService
 
     public static function update($pageid, $title, $showIndex)
     {
-        $nt=time()*1000;
-        $sql = "update page set title='" . $title . "' , status=20, showindex=".$showIndex.",udate=".$nt." where pageid=" . $pageid;
+        $nt = time() * 1000;
+        $sql = "update page set title='" . $title . "' , status=20, showindex=" . $showIndex . ",udate=" . $nt . " where pageid=" . $pageid;
         $mysql = new MySQL();
         $mysql->execute($sql);
         $mysql->closeCon();
